+++ sdk/nodejs/.eslintrc.js
@@ -6,6 +6,7 @@ module.exports = {
     "parser": "@typescript-eslint/parser",
     "parserOptions": {
         "project": "tsconfig.json",
+        "tsconfigRootDir": __dirname,
         "sourceType": "module"
     },
     "plugins": [