+++ sdk/python/lib/pulumi/automation/_local_workspace.py
@@ -302,6 +302,11 @@ class LocalWorkspace(Workspace):
     def install_plugin(self, name: str, version: str, kind: str = "resource") -> None:
         self._run_pulumi_cmd_sync(["plugin", "install", kind, name, version])
 
+    def install_plugin_from_server(self, name: str, version: str, server: str) -> None:
+        self._run_pulumi_cmd_sync(
+            ["plugin", "install", "resource", name, version, "--server", server]
+        )
+
     def remove_plugin(
         self,
         name: Optional[str] = None,